% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_allview.R
\name{dcm_allview}
\alias{dcm_allview}
\title{DICOM Allview}
\arguments{
\item{directory}{path for directory of interest. Defaults to working directory}
}
\description{
Generates images for all DICOM files in the directory
}
\examples{
dcm_allview()

}
