% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_metadata.R
\name{dcm_metadata}
\alias{dcm_metadata}
\title{DICOM Metadata}
\usage{
dcm_metadata(directory = ".")
}
\arguments{
\item{directory}{path for directory of interest. Defaults to working directory}
}
\value{
hdr.data dataframe containing all DICOM header metadata
}
\description{
Extracts DICOM metadata in data frame for all .dcm files in directory
}
\examples{
dcm_metadata()

}
