% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_select.R
\name{dcm_select}
\alias{dcm_select}
\title{DICOM Select}
\usage{
dcm_select(directory = ".")
}
\arguments{
\item{directory}{path for directory of interest. Defaults to working directory}
}
\value{
hdr.data dataframe containing all DICOM header metadata and user evaluations
}
\description{
Displays images and metrics for all .dcm images in the target directory for user evaluation
}
\examples{
dcm_select()

}
